refactor(fb15b): extract bedrconverse message-shaping to utils
Move mapDiracToolsToBedrockToolConfig/processImageContent/
applyCacheControlToMessages + cache-point type into converse-utils.
bedrock 900->790. 38 tests pass.
